+++ mobile-sdk/src/main/kotlin/com/paydock/core/presentation/util/WidgetLoadingDelegate.kt
@@ -1,0 +1,6 @@
+package com.paydock.core.presentation.util
+
+public interface WidgetLoadingDelegate {
+    public fun widgetLoadingDidStart()
+    public fun widgetLoadingDidFinish()
+}