+++ mobile-sdk/src/main/kotlin/com/paydock/designsystems/components/link/HyperlinkText.kt
@@ -18,11 +18,16 @@ import com.paydock.designsystems.theme.Theme
 /**
  * A composable that displays text as a clickable hyperlink.
  *
+ * @param enabled Controls the enabled state of this hyperlink.
  * @param text The text to be displayed as a hyperlink.
  * @param url The URL to be opened when the hyperlink is clicked.
  */
 @Composable
-internal fun HyperlinkText(text: String, url: String) {
+internal fun HyperlinkText(
+    enabled: Boolean = true,
+    text: String,
+    url: String
+) {
     val context = LocalContext.current
     val hyperlinkColor = Theme.colors.primary
     val hyperlinkStyle = SpanStyle(
@@ -42,9 +47,11 @@ internal fun HyperlinkText(text: String, url: String) {
         text = annotatedString,
         modifier = Modifier
             .clickable {
-                val uri = Uri.parse(url)
-                val intent = Intent(Intent.ACTION_VIEW, uri)
-                context.startActivity(intent)
+                if (enabled) {
+                    val uri = Uri.parse(url)
+                    val intent = Intent(Intent.ACTION_VIEW, uri)
+                    context.startActivity(intent)
+                }
             },
         style = Theme.typography.body1
     )