+++ mobile-sdk/src/main/kotlin/com/paydock/feature/paypal/checkout/presentation/PayPalWidget.kt
@@ -6,25 +6,35 @@ import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.testTag
 import com.paydock.R
 import com.paydock.core.MobileSDKConstants
 import com.paydock.core.domain.error.exceptions.PayPalException
+import com.paydock.core.presentation.extensions.alpha40
 import com.paydock.core.presentation.extensions.getMessageExtra
 import com.paydock.core.presentation.extensions.getStatusExtra
 import com.paydock.core.presentation.ui.preview.LightDarkPreview
+import com.paydock.core.presentation.util.WidgetLoadingDelegate
+import com.paydock.designsystems.components.loader.SdkButtonLoader
+import com.paydock.designsystems.theme.PayPal
 import com.paydock.designsystems.theme.SdkTheme
+import com.paydock.designsystems.theme.Theme
 import com.paydock.feature.charge.domain.model.integration.ChargeResponse
 import com.paydock.feature.paypal.checkout.presentation.components.PayPalButton
 import com.paydock.feature.paypal.checkout.presentation.state.PayPalCheckoutUIState
@@ -40,6 +50,8 @@ import org.koin.androidx.compose.koinViewModel
  * A Composable for handling PayPal payments and related interactions.
  *
  * @param modifier Modifier for customizing the appearance and behavior of the Composable.
+ * @param enabled Controls the enabled state of this Widget. When false,
+ * this component will not respond to user input, and it will appear visually disabled.
  * @param token A callback to obtain the wallet token asynchronously.
  * @param requestShipping Flag passed to determine if PayPal will ask the user for their shipping address.
  * @param completion A callback to handle the Wallet Charge result.
@@ -48,8 +60,10 @@ import org.koin.androidx.compose.koinViewModel
 @Composable
 fun PayPalWidget(
     modifier: Modifier = Modifier,
+    enabled: Boolean = true,
     token: (onTokenReceived: (String) -> Unit) -> Unit,
     requestShipping: Boolean = true,
+    loadingDelegate: WidgetLoadingDelegate? = null,
     completion: (Result<ChargeResponse>) -> Unit,
 ) {
     val context = LocalContext.current
@@ -61,8 +75,6 @@ fun PayPalWidget(
 
     val scope = rememberCoroutineScope()
 
-    var isPayPalFlowLaunched by remember { mutableStateOf(false) }
-
     // ActivityResultLauncher for handling payment resolution
     val resolvePaymentForResult = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.StartActivityForResult()
@@ -72,18 +84,18 @@ fun PayPalWidget(
 
     LaunchedEffect(uiState) {
         when (val state = uiState) {
-            is PayPalCheckoutUIState.Idle,
-            is PayPalCheckoutUIState.Loading -> Unit
+            is PayPalCheckoutUIState.Idle -> Unit
+            is PayPalCheckoutUIState.Loading -> {
+                loadingDelegate?.widgetLoadingDidStart()
+            }
             is PayPalCheckoutUIState.LaunchIntent -> {
-                if (!isPayPalFlowLaunched) {
-                    val (callbackData) = state
-                    callbackData.callbackUrl?.let { callbackUrl ->
-                        if (callbackUrl.isNotBlank()) {
-                            val intent = Intent(context, PayPalWebActivity::class.java)
-                                .putCallbackUrlExtra(callbackUrl)
-                            resolvePaymentForResult.launch(intent)
-                            isPayPalFlowLaunched = true
-                        }
+                loadingDelegate?.widgetLoadingDidFinish()
+                val (callbackData) = state
+                callbackData.callbackUrl?.let { callbackUrl ->
+                    if (callbackUrl.isNotBlank()) {
+                        val intent = Intent(context, PayPalWebActivity::class.java)
+                            .putCallbackUrlExtra(callbackUrl)
+                        resolvePaymentForResult.launch(intent)
                     }
                 }
             }
@@ -94,12 +106,14 @@ fun PayPalWidget(
             }
 
             is PayPalCheckoutUIState.Success -> {
+                loadingDelegate?.widgetLoadingDidFinish()
                 completion(Result.success(state.chargeData))
                 // This ensures that we clear the state so it's not reused
                 viewModel.resetResultState()
             }
 
             is PayPalCheckoutUIState.Error -> {
+                loadingDelegate?.widgetLoadingDidFinish()
                 completion(Result.failure(state.exception as Throwable))
                 // This ensures that we clear the state so it's not reused
                 viewModel.resetResultState()
@@ -109,16 +123,46 @@ fun PayPalWidget(
 
     SdkTheme {
         Box(modifier = modifier, contentAlignment = Alignment.Center) {
-            // Button to initiate PayPal transaction
-            PayPalButton(isEnabled = uiState !is PayPalCheckoutUIState.Loading, isLoading = uiState is PayPalCheckoutUIState.Loading) {
-                // Use the callback to obtain the token asynchronously
-                token { obtainedToken ->
-                    scope.launch {
-                        viewModel.setWalletToken(obtainedToken)
-                        viewModel.getWalletCallback(
-                            walletToken = obtainedToken,
-                            requestShipping = requestShipping
-                        )
+            if (loadingDelegate == null && uiState is PayPalCheckoutUIState.Loading) {
+                Button(
+                    onClick = {},
+                    modifier = Modifier
+                        .testTag("loadingPayPalButton")
+                        .fillMaxWidth()
+                        .height(Theme.dimensions.buttonHeight),
+                    enabled = true,
+                    colors = ButtonDefaults.buttonColors(
+                        containerColor = PayPal,
+                        disabledContainerColor = PayPal.alpha40
+                    ),
+                    shape = Theme.buttonShapes.small
+                ) {
+                    Row(
+                        horizontalArrangement = Arrangement.spacedBy(
+                            Theme.dimensions.buttonSpacing,
+                            Alignment.CenterHorizontally
+                        ),
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        SdkButtonLoader(color = Color.Black)
+                    }
+                }
+            } else {
+                // Button to initiate PayPal transaction
+                PayPalButton(
+                    isEnabled = uiState !is PayPalCheckoutUIState.Loading && enabled,
+                    isLoading = false
+                ) {
+                    // Use the callback to obtain the token asynchronously
+                    viewModel.setLoadingState()
+                    token { obtainedToken ->
+                        scope.launch {
+                            viewModel.setWalletToken(obtainedToken)
+                            viewModel.getWalletCallback(
+                                walletToken = obtainedToken,
+                                requestShipping = requestShipping
+                            )
+                        }
                     }
                 }
             }